backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
motor==3.4.0
pymongo==4.6.2
python-dotenv==1.0.1
pydantic-settings==2.2.1
